pacstrap: Rework action to let user define their repositories
Before this commit, we hard-coded the 3 repositories 'core', 'extra',
and 'community', and we only allowed the user define the mirror and
mirror layout for those.

This is not really a good idea, as none of these repositories are
required to exist, or even to be named that way. These are just
conventions. An Archlinux derivative could as well decide to have their
repositories named 'main', 'contrib' and 'non-free', for example.

Hence it seems more flexible to have a 'repositories' property instead,
and let user define the repository name and the repository server. This
way, there's no need for a mirror and mirror-layout properties.

With this change, a basic recipe to build an Archlinux OS can be:

    architecture: amd64

    actions:
      - action: pacstrap
        repositories:
          - name: core
            server: http://mirrors.kernel.org/archlinux/$repo/os/$arch
          - name: extra
            server: http://mirrors.kernel.org/archlinux/$repo/os/$arch

      - action: pack
        compression: gz
        file: recipe.tar.gz

More complicated, a recipe that allows to build both ArchLinux or
Manjaro can be defined like that:

    {{  $arch := or .arch "amd64" -}}
    {{- $distrib := or .distrib "archlinux" -}}
    {{- $mirror := or .mirror "http://mirrors.kernel.org/archlinux" -}}
    {{- $layout := or .layout "$repo/os/$arch" -}}

    {{- if eq $distrib "manjaro" -}}
    {{- $mirror = or .mirror "http://repo.manjaro.org.uk" -}}
    {{- $layout = or .layout "stable/$repo/$arch" -}}
    {{- end -}}

    architecture: {{ $arch }}

    actions:
      - action: pacstrap
        repositories:
          - name: core
            server: {{ $mirror }}/{{ $layout }}
          - name: extra
            server: {{ $mirror }}/{{ $layout }}
          - name: community
            server: {{ $mirror }}/{{ $layout }}

      - action: pack
        compression: gz
        file: {{ $distrib }}-base.tar.gz

And it can be invoked in the following ways:

    # build archlinux:
    debos recipe.yaml

    # build manjaro:
    debos -t distrib:manjaro recipe.yaml

Introducing the 'repository' property has the consequence of removing
the 'mirror' property, which can't make anymore. It means that debos
doesn't try to handle the mirror list anymore, it's left to the user.

By default, ArchLinux leaves you with no mirror configured, so pacman
fails until you manually enable a mirror in /etc/pacman.d/mirrolist.

Manjaro, on the other hand, leaves you with a mirrorlist where all the
mirrors are enabled, and possibly NOT sorted. Watch the pacstrap logs
for the messages:

    Pacstrap | ::INFO Internet connection appears to be down
    Pacstrap | ::INFO Mirror ranking is not available
    Pacstrap | ::INFO Mirror list is generated using random method
    Pacstrap | ::INFO Writing mirror list

One last word.

Remember that these repositories are only used to create the pacman.conf
file used during the pacstrap action. pacstrap itself will install a
default pacman config and a default mirror list in the bootstrapped
system, hence subsequent calls to the `pacman` action will use those.

So if you want to use a particular mirror, defining it in the pacstrap
action is not enough, you will also need to enable it in the
bootstrapped system.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>